+++ source/include/Ostap/BSpline.h
@@ -1536,6 +1536,18 @@ namespace Ostap
     namespace Interpolation
     {
       // ======================================================================
+      class Table ;
+      // ======================================================================
+      /** define parameters for the interpolation spline 
+       *  @param data (INPUT)  table of data 
+       *  @param bs   (UPDATE) the spline 
+       *  @return status code 
+       */
+      Ostap::StatusCode
+      bspline 
+      ( const Ostap::Math::Interpolation::Table& data ,
+        Ostap::Math::BSpline&                    bs   ) ;      
+      // ======================================================================
       /** define parameters for the interpolation spline 
        *  @param xy (INPUT)   vector of data 
        *  @param bs (UPDATE) the spline 
@@ -1601,7 +1613,7 @@ namespace Ostap
         const std::vector<double>& knots     , 
         const unsigned short       order = 3 ) 
       {  
-        // constuct spline  
+        // construct spline  
         Ostap::Math::BSpline      bs ( knots , order ) ;
         // get greville abscissas
         const std::vector<double> ga ( bs.greville_abscissas() ) ;